pyproject.toml
[project]
name = "after-visit-summary-agent"
version = "0.1.0"
description = "AI agent for generating after-visit summaries from EHR and transcripts"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "python-dotenv",
    "scikit-learn",
    "nltk",
    "numpy",
    "openai",
    "google-generativeai",
    "tqdm",
    "requests",
    "genai",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "httpx"
]
